refactor: 컨트롤러 테스트 공통 설정을 추상클래스로 분리
