+++ src/test/java/imhong/dowith/common/ControllerTest.java
@@ -1,0 +1,23 @@
+package imhong.dowith.common;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+@SuppressWarnings("NonAsciiCharacters")
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+public abstract class ControllerTest {
+
+    @LocalServerPort
+    private int port;
+
+    @BeforeEach
+    public void setPort() {
+        RestAssured.port = port;
+    }
+}
